+++ react/components/SKUSelector/constants/CustomClasses.js
@@ -1,0 +1,5 @@
+export default {
+  SKU_SELECTOR: 'vtex-sku-selector',
+  SELECTOR_MANAGER: 'vtex-selector-manager',
+  SELECTOR__ITEM: 'vtex-selector__item',
+}