+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added - 2018-05-22
+- Initial version of SKU selector component.
+
 ### Fixed
 - Fix bug in the url when search something in the product page and click enter.
 - Footer component padding
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `TechnicalSpecifications` component.
 
 ## [1.1.0] - 2018-05-21
+
 ### Changed
 - Delayed queries on _BuyButton_ Component.
 - Add error treatment on _BuyButton_ Component.